move RankStore to /logic
